+++ packages/cbioportal-clinical-timeline/src/Timeline.tsx
@@ -26,7 +26,6 @@ import './timeline.scss';
 import { DownloadControls } from 'cbioportal-frontend-commons';
 import CustomTrack, { CustomTrackSpecification } from './CustomTrack';
 import CustomTrackHeader from './CustomTrackHeader';
-import { TIMELINE_TRACK_HEIGHT } from './TimelineTrack';
 
 interface ITimelineProps {
     store: TimelineStore;
