+++ packages/cbioportal-clinical-timeline/src/TimelineTracks.tsx
@@ -1,15 +1,8 @@
-import {
-    TickIntervalEnum,
-    TimelineEvent,
-    TimelineTick,
-    TimelineTrackSpecification,
-} from './types';
-import { TIMELINE_TRACK_HEIGHT, TimelineTrack } from './TimelineTrack';
+import { TimelineTrack } from './TimelineTrack';
 import React, { useCallback, useState } from 'react';
 import { TimelineStore } from './TimelineStore';
 import _ from 'lodash';
 import { observer } from 'mobx-react';
-import $ from 'jquery';
 import { Portal } from 'react-overlays/lib';
 import { Popover } from 'react-bootstrap';
 import { flattenTracks, sortNestedTracks } from './lib/helpers';
