+++ packages/cbioportal-clinical-timeline/src/index.tsx
@@ -4,8 +4,6 @@ export { TimelineStore } from './TimelineStore';
 
 export { default as Timeline } from './Timeline';
 
-export { TIMELINE_TRACK_HEIGHT } from './TimelineTrack';
-
 export * from './lib/helpers';
 
 export * from './types';
