+++ packages/cbioportal-clinical-timeline/src/lib/helpers.ts
@@ -2,12 +2,25 @@ import {
     TimelineEvent,
     TimelineTick,
     TimelineTrackSpecification,
+    TimelineTrackType,
 } from '../types';
 import { intersect } from './intersect';
 import _ from 'lodash';
-import { getTrackHeight } from '../TimelineTrack';
 
+export const TIMELINE_TRACK_HEIGHT = 20;
+export const TIMELINE_LINE_CHART_TRACK_HEIGHT = 100; // TODO: dynamic?
 export const REMOVE_FOR_DOWNLOAD_CLASSNAME = 'tl-remove-for-download';
+
+export function getTrackHeight(track: TimelineTrackSpecification) {
+    switch (track.trackType) {
+        case TimelineTrackType.LINE_CHART:
+            return TIMELINE_LINE_CHART_TRACK_HEIGHT;
+        case TimelineTrackType.DEFAULT:
+        case undefined:
+        default:
+            return TIMELINE_TRACK_HEIGHT;
+    }
+}
 
 export function getAttributeValue(name: string, event: TimelineEvent) {
     const att = _.at(event as any, ['event.attributes']);
